requirements.txt
twikit
pymongo
python-dotenv
sentence-transformers
kafka-python
qdrant-client
numpy
fastapi
uvicorn
tenacity